xps9700: prevent X server from loading nvidia module
This is a potential fix to the Xorg hang problem when resuming from
hybrid-sleep.

The symptom is that suspend and hibernate work fine, but not hybrid-sleep.
When wake up, Xorg takes 100% CPU with log entry about nvidia module
failes to find object.
The OS is still fully functional because SSH session fully working.